[refactor] Client 분리
