+++ src/main/java/ladder/domain/HowToConnect.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public interface HowToConnect {
     boolean needToConnect(Boolean leftSideCondition);
