+++ src/main/java/ladder/domain/User.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 
 import ladder.ladderexceptions.InvalidUserNameException;
