+++ src/test/java/ladder/domain/LadderTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 
 import ladder.ladderexceptions.InvalidLadderHeightException;
